Updates to resolve issues with retrieve multiple requests

+++ src/main/java/com/impinj/itemsense/client/coordinator/zonemap/Point.java
@@ -4,7 +4,7 @@ package com.impinj.itemsense.client.coordinator.zonemap;
  * Created by jcombopi on 1/25/16.
  */
 public class Point {
-    private int x;
-    private int y;
-    private int z;
+    private double x;
+    private double y;
+    private double z;
 }
